requirements.txt
numpy
pandas
matplotlib
seaborn
streamlit
tensorflow
scikit-learn
ipykernel
Pillow